ci(lint): fix errors reported by pylint
